Refactor deploy and extract DoSSA to pkg
